Introduce Kotlin-specific `Usage` values in addition to Java ones
We need this to be able to publish a Kotlin library with several
variants and consume it in a JVM project that is unaware of Kotlin.

Our initial scheme would cause a conflict of multiple `java-api` and/or
`java-runtime-jars` variants generated by multiple Kotlin targets.

This commit fixes that by leaving the Java `Usage` values only
for the JVM targets and introducing separate Kotlin-specific Usage
values.

If the Kotlin plugin is applied in a project, the project becomes aware
of the Kotlin Usage constants (through the input configuration
attributes) and is therefore able to consume those artifacts.